docs: Convert ASCII diagrams to Mermaid and refactor daemon/operator code
Documentation:
- Convert all ASCII box diagrams to Mermaid flowcharts/sequence diagrams
- Restructure main README with problem-first opening and user journey sections
- Add unified deploy/ directory with production Kustomize manifests
- Add monitoring-alerting.md guide with Prometheus metrics and Grafana setup
- Fix technology versions in panoptes-eye README (Next.js 16, React 19)
- Add deprecation notices to proto/v1 READMEs

Code refactoring:
- Extract helper functions in argusd (resolve_pids, build_config, register_watches)
- Extract helper functions in janusd (resolve_pids, build_config, spawn_event_loop)
- Add helpers.go with shared utilities for both operators
- Simplify controller reconciliation logic with extracted methods
- Update CRD schemas with improved validation

Files converted to Mermaid:
- docs/security/threat-model.md (trust boundaries)
- docs/guides/multi-cluster.md (central observability)
- docs/security/cryptographic-guarantees.md (TLS, key management)
- docs/security/network-monitoring-roadmap.md (Hermes architecture)
- daemons/*/README.md (Rust architecture diagrams)
- daemons/common/EBPF.md (eBPF architecture, filtering flow)
- tools/*/README.md (pod startup sequence diagrams)
- operators/*/README.md (architecture, webhook flow)
- ui/panoptes-eye/README.md (Next.js architecture)

+++ daemons/janusd/src/audit.rs
@@ -302,7 +302,7 @@ impl NetlinkAuditLogger {
         // nlmsg_seq (u32)
         buffer[8..12].copy_from_slice(&seq.to_ne_bytes());
         // nlmsg_pid (u32) - our PID
-        buffer[12..16].copy_from_slice(&(std::process::id() as u32).to_ne_bytes());
+        buffer[12..16].copy_from_slice(&std::process::id().to_ne_bytes());
 
         // Copy message payload
         buffer[NLMSG_HDRLEN..NLMSG_HDRLEN + msg_bytes.len()].copy_from_slice(msg_bytes);
